+++ src/components/CampaignMap.tsx
@@ -2,17 +2,15 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 
 import Preloader from './Preloader';
+import Map from './map/Map';
+import MapImageLayer from './map/MapImageLayer';
+import MapImageTextLayer from './map/MapImageTextLayer';
+import RegionMarkerLayer from './map/RegionMarkerLayer';
+import RegionAreaLayer from './map/RegionAreaLayer';
+import MapEventListener from './map/MapEventListener';
+import MapCenterButton from './map/MapCenterButton';
 
-
-import Map from './leaflet/Map';
-import MapListener from './leaflet/MapListener';
-import ImageLayer from './leaflet/ImageLayer';
-import ImageTextLayer from './leaflet/ImageTextLayer';
-import RegionMarkers from './leaflet/RegionMarkers';
-import RegionSvgLayer from './leaflet/RegionSvgLayer';
-import MapCenterButton from './leaflet/MapCenterButton';
-
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(() => ({
   map: {
     position: 'relative',
     flex: 1,
@@ -31,11 +29,11 @@ const CampaignMap = (props: CampaignMapProps) => {
     <div className={classes.map} key={campaign.key}>
       <Preloader assets={[campaign.map, campaign.mapText]}>
         <Map campaign={campaign}>
-          <ImageLayer />
-          <ImageTextLayer />
-          <RegionSvgLayer />
-          <RegionMarkers />
-          <MapListener />
+          <MapImageLayer />
+          <MapImageTextLayer />
+          <RegionAreaLayer />
+          <RegionMarkerLayer />
+          <MapEventListener />
           <MapCenterButton />
         </Map>
       </Preloader>
