+++ src/components/map/MapCenterButton.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react';
 import { Fab } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { FilterCenterFocus } from '@material-ui/icons';
-import { useMapContext } from './map-context';
+import { useMapContext } from './map';
 
 const useStyles = makeStyles((theme) => ({
   fab: {
